requirements.dev.txt
ruff>=0.13, <0.14
textual-dev>=1.7, <1.8
build
